move some of the substitution framework into kinding to avoid repetition (in the face of circular dependencies)
