+++ frontend/src/Language/Granule/Checker/Coeffects.hs
@@ -6,7 +6,6 @@ module Language.Granule.Checker.Coeffects where
 import Language.Granule.Checker.Monad
 import Language.Granule.Checker.SubstitutionContexts
 import Language.Granule.Checker.Kinding
-import Language.Granule.Checker.Substitution
 import Language.Granule.Context
 
 import Language.Granule.Syntax.Identifiers
